+++ kubejs/server_scripts/recipes/processing_lines/perfect_electrum.js
@@ -1,7 +1,7 @@
-import { recipe_chem_bath, recipe_teus_laser } from "../../00_util/recipeUtils"
+
 
 ServerEvents.recipes((event) => {
-    global.recipe_chem_bath(event, "gtceu:perfected_electrum_base_dust", "4x gtceu:graphene_dust", "gtceu:refined_fluxed_electrum 288", 1024, 30)
+    global.recipe_chem_bath(event, "gtceu:perfected_electrum_base_dust", "4x gtceu:titanite_dust", "gtceu:refined_fluxed_electrum 288", 1024, 30)
 
     global.recipe_teus_laser(
         event,