+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 ### Feature Additions
 + Added 4 new steam multiblocks (Ore washer, Centrifuge, Forge Hammer, Compressor)
 + Basic helper JEI recipe integration
++ Added Multiblock Water Collector
 
 ### Balance/Recipe Changes
 + Gregified some of Forestry, mainly through recipe alterations in the base components.
@@ -42,6 +43,8 @@
 + Changed IaF recipes
 + Added recipes for Helper Assemblers
 + Kitchen sink no longer has infinite water
++ Added recipe for Magma cube pet
++ Inventory pets no longer need food
 
 ### Mod Additions
 + Added Aether